plotly.js: Move trace modules to the correct location
https://github.com/plotly/plotly.js/tree/v2.35.2/lib
