+++ types/plotly.js/lib/candlestick.d.ts
