+++ types/plotly.js/lib/ohlc.d.ts
