+++ types/plotly.js/lib/scatter.d.ts
@@ -1,4 +1,4 @@
-import { Color } from "../../index";
+import { Color } from "../";
 
 export interface ScatterSelectedMarker {
     marker: Partial<{
